refactor(api): extract the D1 analytics + edge-cache cluster into request-handlers/analytics.mjs
Moves the five D1-backed analytics handlers (bulk health trends, per-subnet
trends, percentiles, incidents, global incidents) together with withEdgeCache
and the full D1-fallback machinery (d1FallbackGeneration counter, the two
WeakSets, markD1FallbackRows/hasD1FallbackRows/markD1FallbackResponse) plus the
shared d1All/d1Runner read path out of workers/api.mjs into a new
workers/request-handlers/analytics.mjs.

Co-locating the cache guard with the handlers it protects keeps the
await/WeakSet contract (the #1760 poison-protection) reviewable in one file:
the fallback counter is mutated where d1All catches a D1 error and read inside
withEdgeCache, so they must share one module's state — they now do.

api.mjs imports d1All/d1Runner/withEdgeCache and the handlers back and keeps the
router. The one api.mjs-local dependency (readHealthMetaKv) is injected once via
configureAnalytics() so the two modules never import each other.

Refs #1763
